formatter/pyproject.toml
[project]
name = "formatter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.64.0",
    "python-dotenv>=1.2.1",
    "flask>=3.0.0",
    "flask-login>=0.6.3",
    "requests>=2.31.0",
]
